Move ProfilePhoto to shared instead of accounts.
