+++ frontend/src/routes/posts/index.ts
@@ -1,8 +1,7 @@
-import CreatePost from "./CreatePost";
-import ViewFeed from "./ViewFeed";
-import ViewPost from "./ViewPost";
-import ShowPost from "./ShowPost";
-import ViewUserFeed from "./ViewUserFeed";
-import PostGrid from "./PostGrid";
+import CreatePost from './CreatePost';
+import ViewFeed from './ViewFeed';
+import ViewPost from './ViewPost';
+import ShowPost from './ShowPost';
+import ViewUserFeed from './ViewUserFeed';
 
 export {CreatePost, ViewFeed, ViewPost, ShowPost, ViewUserFeed};
